Update parser structure
