+++ src/lib.rs
@@ -2,7 +2,7 @@
 #![feature(bool_to_option)]
 #[macro_use]
 pub mod common_ffi;
-pub mod ast;
 pub mod dictionary;
 pub mod math;
 pub mod namespaces;
+pub mod parser;
